fix không tiếp cận được boss
